Make DeckOfCards generate Json objects
Replaced logic for generateDeck.
Now generates an array of Json objects, storing the card, cards string-value, cards int-value and cards symbol in each object.
This made it possible to replace the logic for retrieving a cards symbol, reducing code and unessescary calls.

Now uses the same logic for api/json calls as for the normal routes, reducing code and complexity.