Add new implementation of DeckOfCards
Added new methods to support a new implementation of GenerateDeck => generateGraphicDeck

GenerateDeck was a bad implementation where each element in the array of cards (jsonCards) where a dictionary with
keys and values (card:GraphicCard objekt, value:int value, stringvalue: value as string, symbol:utf-8 symbol).
This made it more complicated then nessescary.

Implementation is still backwards compatible.

New implementation with generateGraphicDeck insted create a complete deck of cards as GraphicCard objects.
This makes it more clear that the array is just an array of GraphicCard objects and values as symbol, stringvalue
int value etc is stored in the object itself.